requirements.txt
Flask==3.0.3
pandas==2.2.2
scikit-learn==1.2.2
google-cloud-translate==3.15.3
joblib==1.4.0
nltk==3.8.1
fasttext-wheel==0.9.2
numpy==1.23.5
scipy==1.10.1
IPython==8.18.1
moviepy==1.0.3